[armel/marvell,sh4] linux-image: Recommend apparmor, like all other configs
The "recommends" field set in the [image] section for these
configurations overrode the field at the top level.  We want
gencontrol.py to concatenate the relations in this section at all
levels.

The ConfigCore.get_merge method supports doing this, but only with
list fields So we need to specify in the config schema that these
fields are comma-separated lists.