[arm64] udeb: kernel-image: Include cros_ec_spi and SPI drivers
The cros_ec multifunction device provides the keyboard services on some
ChromeOS devices, but requires a bus to be enabled to communicate with
it. On Samsung Chromebook Plus (v1), including spi-rockchip and
cros_ec_spi are enough. A recent change in initramfs-tools included all
SPI drivers, so include them here as well.

Relevant:
  https://salsa.debian.org/kernel-team/initramfs-tools/commit/797e5fed

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>