Drop unnecessary changes from "libbpf: add SONAME to shared object"
It's not necessary to delete the definitions of the variables that
become unused.  Nor is it necessary to move the definition of
LIBBPF_VERSION before LIB_FILES, because the latter is defined
as recursively expanded (i.e. its variable references are not
immediately expanded).

This makes the actual change we're making clearer, and should
reduce the future work to maintain this patch.