gencontrol: Generalise substitution of debhelper config template
Currently we have two copies of the _substitute_file() method, and
lots of somewhat similar invocations of it since different binary
packages need different sets of debhelper configs.

We already name the templates in a fairly consistent way, so we can
replace this with a loop that tries to expand all possible templates
for a package and ignores those that are missing.  Add a method for
this to the Gencontrol base class and use it in both subclasses
where we prevously used _substitute_file().