crypto: Explicitly enable ASYMMETRIC_KEY_TYPE, ASYMMETRIC_PUBLIC_KEY_SUBTYPE
These are selected by SYSTEM_DATA_VERIFICATION, which is in turn
selected by MODULE_SIG and CFG80211_REQUIRE_SIGNED_REGDB.  So far as I
can see, one or both of MODULE_SIG and CFG80211_REQUIRE_SIGNED_REGDB
are enabled in all configs except m68k.

It would be nice if we could avoid this on armel/marvell, but I don't
think that's going to be practical.