debian/bin/gencontrol_signed.py: Generate valid versions in a linux binNMU
After a binNMU of linux, we will trigger the signing service to attempt
source uploads of linux-signed-*.  But their version strings will end in
"+b<n>", which will be rejected by dpkg-source as invalid.

Replace the "+" with a "." in the linux-signed-* source version.  The
binary versions will still include "+b<n>", which is weird for a
non-binNMU but should still be accepted.