[x86] Add support for mitigation of SRBDS (CVE-2020-0543)
Apply the current version of the backport to 5.6.

Note that this currently requires an ABI bump.  Depending on what
other changes are included in the next upload, we may be able to avoid
that by dropping "x86/cpu: Add a steppings field to struct x86_cpu_id"
and changing VULNBL_INTEL_STEPPINGS() in arch/x86/kernel/cpu/common.c
to use VULNWL_INTEL().