Don't run dh_strip on vmlinuz
The debian kernel build switched to use debhelper
compatibility level 12 with this commit:

commit b02f531e01769fce339e477dc00026cb12971464
Author: Ben Hutchings <ben@decadent.org.uk>
    debhelper started complaining about level 9, so it's time to upgrade
    again.

This change broke the kernel build on the hppa architecture.
Since compat level 12, dh_strip runs the "strip" command on any vmlinuz*
files, which wasn't the case before.
On hppa the 64-bit build needs to run "hppa64-linux-strip" instead of
"strip".

Fix it by adding "vmlinuz" to the exclude list of dh_strip, and thus
behave as compat level 9 did before.

Tested on hppa and x86-64.