Fix conversion of meta-package doc directories to symlinks (Closes: #942861)
'dpkg-maintscript-helper dir_to_symlink' takes a prior-version
parameter, which is the latest version that needs conversion when we
upgrade from it.  I interpreted this is as being the latest version
that shipped the directory, but this is wrong.

It really needs to be the latest version that shipped the directory
*or* failed to do the conversion.  So set prior-version to the current
version in experimental.