[armhf] Enable NVMEM_IMX_OCOTP
On iMX6Q boards, when probing cpufreq, it tries to read efuse settings,
however, as NVMEM_IMX_OCOTP is not enabled this can't be done. In this
situation the probe fails with -EPROBE_DEFER, and cpufreq is not
available.

To avoid all the mentioned issues, while these issues are fixed in the
upstream kernel, enable NVMEM_IMX_OCOTP as module.

Signed-off-by: Walter Lozano <walter.lozano@collabora.com>