[amd64/cloud-amd64] Generalise the description of supported platforms
The cloud-amd64 flavour is being used successfully in OpenStack-based
clouds, and should work in a typical libvirt/KVM/QEMU VM.  Change the
description to say "cloud platforms including …", to avoid giving the
impression that it's limited to specific private clouds.

While we're at it, also move Azure ahead of GCE as it's more widely
used.