control.image.meta.in: set Provides: virtualbox-guest-modules
With virtualbox guest modules now being part of Debian's kernel, it makes sense to have
the kernel package satisfy the virtual package slot, "virtualbox-guest-modules",
that is required by the virtualbox-guest* meta package. This, notably, is what's
already done by Ubuntu's kernel as well as Arch's, and other distro
kernels are likely to follow. Note that this is a somewhat temporary
arrangement -- we only need this while we're keeping virtualbox-guest-dkms
around to support kernels that don't come with wireguard built-in.
Over time that number will dwindle, at which point we can get rid of
virtualbox-guest-dkms as well as reverting this commit. In other words, adding
this Provides: line is a temporary stop-gap solution while wireguard
transitions from being an out-of-tree module to an in-tree module.

(taken from commit d79ff3533ee8b028a9cef6a579270c94670141ce)

Signed-off-by: Gianfranco Costamagna <locutusofborg@debian.org>