debian/tests/python: pycodestyle: Increase max-line-length to 100
Set the maximum allowed line length to 100 as the default of 79 is too
limiting small. Raising the allowed line length to 100 seems acceptable
enough.

This addresses autopkgtest failure for the python test since the
refactoring changes for debian/bin/gencontrol.py to factor out the
siable flags for installer modules, debug info and signed code.

Fixes: b312397e81e5 ("Factor out installer modules disable flag")
Fixes: a0dd8aecc3ee ("Factor out debug info disable flag")
Fixes: 5b3406213441 ("Add flag to disable signed code")
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>