Enable USB_NET_AQC111 as module
This is a driver for several USB to 5GBase-T ethernet adapters
in the wild.

Previous kernels used cdc_ether for these devices, which was already
enabled.
For v5.4+ in particular, this is required because changes in the cdc_ether
and this driver meant that the QNAP device and others are blocklisted from
cdc_ether as the new acq11 driver is intended to handle AQC111 chipset
devices from now on.

Closes: #968760