Make LEDS_TRIGGER_HEARTBEAT builtin.
The heartbeat trigger is defined as default trigger for LEDs in the
DeviceTree in 281 cases of which 268 belong to ARM dts/dtsi files which
likely means even more boards/SBC. If the module isn't loaded or
builtin, then the trigger gets reset to 'none' which differs from the
intend expressed in the dts[i].

The heartbeat trigger is more versatile and also fulfills (more then)
the same functionality as the panic trigger, which is already builtin.
As blinking of the LED provides important information that the device
has started up, make it builtin on all ARM devices.

Closes: #992184