integrity: Adjust "MODSIGN: do not load mok when secure boot disabled"
Currently the check for efi_enabled(EFI_SECURE_BOOT) is inserted in
the wrong place, but then moved by the following patch "MODSIGN: load
blacklist from MOKx".  Insert it in the right place to start with.

This makes no functional change, or even a change to the completely
patched source.