libcpupower: Update symbols file for changes in 5.13.9-1~exp1
We used to hide the cpupower_read_sysfs() function as it was not
declared in a public header.  However, the cpupower executable now
uses both that and another private function, cpupower_write_sysfs(),
so they are not hidden and must be listed in the symbols file.

cpupower_write_sysfs() was added in 5.11 but the first Debian version
affected was 5.13.9-1~exp1.