d/bin/genorig.py, d/README.source: Add support for remote upstream repos
The common salsa-ci pipeline uses origtargz to download a tarball
from Debian or upstream (in the latter case via uscan).  Ideally
we could do that too, even if we can't currently use the whole of
that pipeline.

Unfortunately, in our configuration uscan will compress the entire
upstream source twice (bug #1003251), so genorig.py is still
preferable.

To avoid duplicating version logic in salsa-ci.yml, extend genorig.py
to support remote repositories.  In case the repository path is
remote, make a temporary shallow clone of it and export from that.