[arm64] Enable Qualcomm SDA845 SoC
Enable components needed for SDA845 SoC
based boards, such as inforce SYS6701 or Thundercomm
rb3 kit.

It allows to boot from UFS storage and to get a serial
console, and a working usb bus.

In order to have the first kernel log, it is required
enable the CONFIG_SERIAL_QCOM_GENI config, as well as
Interconnect, qcom_pdc, clock controllers and pinctrl.

CONFIG_SCSI_UFS_QCOM is added to boot from the flash
storage, and should be included in the initramfs.