d/lib/python/debian_linux: Introduce {Source,Binary}Package classes
The existing Package class has a list of header field names that may
be valid only for source packages, only for binary packages, or valid
for both.  It's also missing a few fields that we use; such fields are
still accepted but without any validation or a specific output
ordering.

Replace it with SourcePackage and BinaryPackage classes, each with the
appropriate subset of field names.

Add read_control_source() to construct SourcePackage, and change
read_control() to construct BinaryPackage.  Use read_control_source()
in the Templates class as appropriate.