d/t/main.control.in: Add Depends on python3-jinja2 for linux-support packages
When adding support for Jinja2 templates in
lib/python/debian_linux/utils.py we added a Builds-Depends accordingly
in the source package. But the utils are installed in the linux-support
binary packages well (and in turn then used e.g. by the firmware-nonfree
packages).

Add the Depends to python3-jinja2 as well to the linux-support binary
packages.

Fixes: 01bd5c810e5e ("Support Jinja2 templates")
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>