Enable limiting of compression threading
The xz compression is currently being configured to use a number of
worker threads equal to the processors on the system being used. This is
generally OK, however in some instances we compile packages on large
machines, with as many as 160 CPU cores. This results in memory
allocation issues:

[ 1137s] find 'linux-source-6.1' \
[ 1137s]     -name '*.pyc' -prune -o \
[ 1137s]     -print0 | \
[ 1137s]     sort -z | \
[ 1137s]     tar --owner=root --group=root --numeric-owner \
[ 1137s]         --no-recursion --null -T - -c | xz -T0 > 'linux-source-6.1.tar.xz'
[ 1137s] xz: (stdin): Cannot allocate memory
[ 1137s] make[2]: *** [debian/rules.real:84:debian/build/linux-source-6.1.tar.xz] Error 1

When we are setting the "parallel" setting in "DEB_BUILD_OPTIONS", use
this to set the number of threads used for compression.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>

[benh: Add a changelog line]