d/b/test-patches: Build a linux-headers-common package as well
test-patches doesn't build a linux-headers-common package, and thus
the linux-headers package it does build is uninstallable.  This
problem was reported in bug #871216 and raised again in bug #1022025.

Currently we directly invoke debian/rules.gen to build the specific
featureset and flavour that we want, and that skips all the common
packages.  Now that debian/config.local is available to make temporary
config changes, a better approach is possible:

- Override the list of featuresets for the target architecture to be
  just the one we want.
- Similarly, override the list of flavours for the target architecture
  and featureset to be just the one we want.  Also override
  default-flavour to ensure it's still valid.
- Enable the nodoc, noudeb, and pkg.linux.nosource build profiles
  since these packages aren't needed and udebs might require building
  a different flavour.
- Build with dpkg-buildpackage -b.

This will actually build several more packages (the exact list
depending on target architecture):

- linux-compiler-<compiler>-<srcarch>
- linux-headers-common-<featureset> for another featureset
- linux-image-<arch>-signed-template
- linux-libc-dev
- linux-support-<abiname>

but these are small, quick to build, and don't add further build-
dependencies.

Closes: #871216