linux-perf: Build C++ code with Debian standard compiler flags
perf now has one C++ source file and we should build it with the
proper compiler flags for hardening etc.

Adding CXXFLAGS to the make command line will override the many
additions that Makefile.config does.  So add support for
EXTRA_CXXFLAGS along the lines of EXTRA_CFLAGS, and use that.