d/rules.real: Fix missing -ffile-prefix-map options for user-space
In commit 1e13025629c6 "Set -fdebug-prefix-map=... in compiler options
for kernel and userland", I set the environment variable
$DEB_BUILD_PATH in debian/rules.real to ensure that out-of-tree builds
would use the correct directory name in the -fdebug-prefix-map
option (now superseded by -ffile-prefix-map).

Unfortunately, by adding a comment on this line, I included trailing
space to the value of this variable, which doesn't get stripped at any
point.  The code in Dpkg::Vendor::Debian module that defines build
flags treats this as unsafe because it's not clear how to escape it
in all cases, and so doesn't emit either of the prefix-map options.

Move the comment onto a separate line.

Closes: #1033663