[arm64,riscv64] drivers/leds/rgb: Enable LEDS_GROUP_MULTICOLOR and LEDS_PWM_MULTICOLOR as modules
This enables the use of multicolor RGB LEDs.
As these modules were added in 2022 and 2023, only enable it on 'modern'
arches like ``arm64`` and ``riscv64``.
If it later turns out that f.e. an ``armhf`` device could also use this,
then support for it could be added later.

The LEDS_PWM_MULTICOLOR module was already enabled for riscv64 in commit
632a2629980f6 ("[riscv64] Enable LEDS_PWM and LEDS_PWM_MULTICOLOR as modules.")