d/copyright: Exclude 'action-ebpf' as that's a binary blob
Upstream commit a79d8ba734bdbd2574ad16dd1b96506e5f642c4a titled
"selftests: tc-testing: remove buildebpf plugin"
removed the/some source code with a precompiled binary.

The secondary commit message is:
"As tdc only tests loading/deleting and anything more complicated is
better left to the ebpf test suite, provide a pre-compiled version of
'action.c' and don't bother compiling it in kselftests or on the fly
at all."

Not sure why upstream appears to be fine with a binary blob (without
source code) in their GPL-2 project.

In Debian's ``src:firmware-nonfree`` package several firmware files are
excluded because it's claimed to be GPL-2, but without the source code.
And that package is meant for the non-free-firmware archive area.

So do the same for this package targetted to the main archive area:
Exclude it via ``Files-Excluded`` in ``debian/copyright``.