drivers/devfreq: Enable PM_DEVFREQ
The "Generic Dynamic Voltage and Frequency Scaling (DVFS) support"
functionality allows the system to change frequencies and voltages based
on the device driver's policy.

The support for this was disabled in the configuration, but effectively it is
enabled on all architectures.
Fix the discrepancy by explicitly enabling PM_DEVFREQ.

As it's now enabled globally, remove the setting for ``armhf``.