[arm64,armhf] drivers/devfreq: Make DEVFREQ_GOV_SIMPLE_ONDEMAND builtin
The DEVFREQ_GOV_SIMPLE_ONDEMAND driver is globally built as a module.
The ARM Mali GPU driver needs a devfreq governor. When it's not loaded,
one would see errors like this (in ``dmesg``):

```
[drm:panfrost_devfreq_init [panfrost]] *ERROR* Couldn't initialize GPU devfreq
```

The result of such an error could be that there is no output on screen.
One way to 'work around' that, is adding DEVFREQ_GOV_SIMPLE_ONDEMAND to
initramfs. But that has the downside that everyone needs to know about
the problem and know that this work around exists.
Making it builtin prevents that error from occuring.
More and more ARM devices do use the ARM Mali GPU drivers, which makes
it even more useful to make those drivers work OOTB.

On 2023-07-20 via a blog post on Collabora's website, it was announced
that ARM was officially supporting the Open Source ARM Mali drivers,
which make it even more compelling to use.

Surprisingly, an arm64 kernel with this module builtin had the exact
same size as one where it was built as a module.

Therefor make DEVFREQ_GOV_SIMPLE_ONDEMAND builtin on both ``arm64`` and
``armhf``.

Link: https://wiki.pine64.org/wiki/ROCKPro64#No_Video_or_GPU_Acceleration_on_Debian
Link: https://www.collabora.com/news-and-blog/news-and-events/a-helping-arm-for-panfrost.html