d/signing_templates/rules.real: Use an intermediate install directory
We currently build linux-image packages first and then use each
package directory as a source directory for kernel-wedge to produce
installer udeb packages.

Since kernel-wedge expects modules to be in the lib/modules
subdirectory of its source directory, this breaks if we run
dh_movetousr in the linux-image build:

https://buildd.debian.org/status/fetch.php?pkg=linux-signed-amd64&arch=amd64&ver=6.7%2B1~exp1&stamp=1704947571&file=log

To avoid this, introduce an intermediate installation directory
matching the one used in the main linux package build, and copy
from this in both build_image and build_installer rules.

Thanks to Bastian for explaining what went wrong.