MIPS64EL: add mips64r6el flavor
The kernel for MIPSr6 has a config CONFIG_MIPSR2_TO_R6_EMULATOR.
With this config is enabled, the R2 userland is supported with the
r6 kernel.

So, with this flavor, we can use R2 userland on r6 CPU such as I6500.
The CI of QEMU project asks for this flavor.

In the upstream code, 'mipsr2emu' kernel option is needed to enable
R2 to R6 emulator.  Let's patch it to enable it always.