d/l/p/debian_linux/firmware.py: Correctly handle multiple licences per section
The firmware WHENCE file has (usually) one section per driver, but in
some cases there are multiple groups of firmware used with the same
driver that are under different licences.

We don't really need to care about representing the per-driver
sections in the WHENCE file, but we do need to associate the licence
information with the right group of files.

To fix this:

- Rename the FirmwareSection class to FirmwareGroup, and update
  comments accordingly.

- When we see another field after a licence field, start a new group.