d/control: Fix profiles for Build-Depends on bison, cpio, flex, xz-utils
All of bison, cpio, flex, and xz are used to build multiple types of
package and as a result they are needed unless all of those packages
are excluded by build profiles.

The build profile syntax supports negation ("!") and conjunction
(whitespace) operators, but not disjunction or parentheses.  This
means we can't directly express a build dependency with profiles
¬(A ∧ B).  Instead we have to transform that to ¬A ∨ ¬B, and then
express the disjunction outside of the profile syntax by duplicating
the dependency:

   D <!A>, D <!B>

In commit b0b32308951b ("Drop deprecated build profile stage1") the
build-dependencies on cpio and xz-utils were incorrectly written as
the conjunction ¬A ∧ ¬B.  For bison and flex build-dependencies, one
of the two build profiles was wrongly omitted.

Commit 5a7ef0c52c35 ("debian/templates: Keep cpio as build-depends for
nodoc build profile") deleted one of the negated build profiles for
cpio, which fixed some but not all cases.

Duplicate all of these build-dependencies to express the build
profiles correctly.

Fixes: b0b32308951b ("Drop deprecated build profile stage1")
Fixes: 5a7ef0c52c35 ("debian/templates: Keep cpio as build-depends ...")