[amd64, arm64] Add d-i support for UEFI and U-Boot HTTP boot
UEFI HTTP boot and the U-Boot equivalent work by making the installation media
available as /dev/pmem0. In order for the installer to be able to mount it, we
need to have pmem modules in the installer initrd. Currently, if the installer
is started via HTTP boot it fails to detect installation media.

https://github.com/tianocore/tianocore.github.io/wiki/HTTP-Boot
https://www.linaro.org/blog/ledge-blogs-uefi-http-and-https-boot-in-u-boot/

Related bug: #879106