[arm64] Enable additional modules for rk356x devices
Based on the v6.12-rockchip-dts64-1 tag I ran my analysis script to
check whether all the modules for rk3566 and rk3568 devices were
enabled/available in the Debian kernel.

I did not analyse 'anberic' and 'powkiddy' devices as those are handheld
gaming consoles and I guess they'd run a custom OS, not Debian, so that
didn't seem useful. No principle reason to exclude them though.

There are 2 modules not (yet) enabled as their driver is supposed to
land in the 6.12 kernel, so they're not yet available. These are:
- drivers/char/hw_random: HW_RANDOM_ROCKCHIP (tristate)
- drivers/net/can/rockchip: CAN_ROCKCHIP_CANFD (tristate)

The ones that got enabled are:
- [arm64] drivers/gpu/drm/panel: Enable DRM_PANEL_SITRONIX_ST7789V as module
- [arm64] drivers/input/touchscreen: Enable TOUCHSCREEN_ST1232 as module
- [arm64] drivers/pci/controller/dwc: Enable PCIE_ROCKCHIP_DW_EP
- [arm64] drivers/regulator: Enable REGULATOR_TPS6286X as module
- [arm64] drivers/rtc: Enable RTC_DRV_PCF85363 as module
- [arm64] drivers/usb/typec: Enable TYPEC_WUSB3801 as module
- [arm64] drivers/video/backlight: Enable BACKLIGHT_LM3630A as module
- [arm64] sound/soc/codecs: Enable SND_SOC_DMIC and SND_SOC_TAS2562 as module