rtla: Switch to out-of-tree build
Now that rtla uses the common tools build infrastructure, it does
support out-of-tree builds and we no longer need to sync the source
across.

We can also stop using a VERSION file because the upstream Makefile
will now be able to run 'make kernelversion' successfully.

Installation from an out-of-tree build currently fails, though, so add
a patch to fix that.  I will send this upstream shortly.