d/config: enable IPE LSM
IPE is a new LSM being introduced in 6.12. Like IMA, it works based on
a policy file that has to be loaded at boot, and is inert otherwise.
In a nutshell, it allows to enable code integrity on a system.

systemd v257 will have the ability to load a policy from /etc/ipe/ just
like it does for IMA.

For more information on the details of IPE and how to write policies,
see:

https://docs.kernel.org/security/ipe.html