[arm64] Enable ARM64_VA_BITS_52
Suggested by ardb in #debian-arm. Currently there is no hardware
available that would use 52-bit virtual addressing yet. But with this
option enabled the kernel is ready for such hardware and on machines
that cannot do that, it falls back to 48 bit addressing which matches
the previous setting. So every machine that works with ARM64_VA_BITS_48
is supposed to also work with ARM64_VA_BITS_52.

Also ARM64_VA_BITS_52 is the default setting for the "Virtual address
space size" choice.