objtool: Fix compiler flags leaking to fixdep in cross-build
After my previous changes to the objtool build, I was able to
cross-build it for arm64, but it seems that it worked by chance.  Now
that the compiler flags for arm64 user-space include
-mbranch-protection=standard, cross-building stopped working again as
this option is wrongly used when compiling the native fixdep.

fixdep (like other simple 'hostprogs') is compiled with options from
KBUILD_HOSTCFLAGS, so we need to clear that variable and not
HOSTCFLAGS.