[arm64] udeb: Add modules for Qualcomm SC7180 Chromebooks
Most of the drivers common to Qualcomm devices are already included in
the installer udebs thanks to other contributors. Include the few clocks
that we need for the SC7180 SoC to boot into the installer. The
gcc-sc7180 driver is currently built-in, but still add it to the list to
match the existing patterns.

Backporting note: The dispcc-sc7180 and gpucc-sc7180 addition to
fb-modules is dropped as they are already in the kernel udeb. This is
similar to commit 6146d64a5c97 ("[arm64] Drop some modules from
fb-modules that are also in the kernel udeb").

Backported from commit 3b9ef43b18326ef914c9a9146a185c5f94b6935e