debian/salsa-ci.yml: Include run of .build-after-script from common pipeline
The Salsa-CI pipeline 66fb5b9bc368 ("Use 'after_script' section to run
commands needed also on failing builds") and a3fbdb16fdf5 ("Save cache
for next run even on failed builds") changed the way commands are run
after the build of the package as the salsaci user and moved some
cleanup in the after_script section.

We miss now this cleanup steps as well where we override the defintions
and need to include the .build-after-script as well.

Since the .build-after-script will include as well the saving of the
ccache directory, remove the explict call.

Link: https://salsa.debian.org/salsa-ci-team/pipeline/-/commit/66fb5b9bc368099fd3d6db0448d182061158d9de
Link: https://salsa.debian.org/salsa-ci-team/pipeline/-/commit/a3fbdb16fdf5a814bd0b2992657cb145fe260139
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>