[arm64] drivers/char/hw_random: Enable HW_RANDOM_ROCKCHIP as module
This enables the HWRNG in the Rockchip rk3566/rk3568 SoC.

There are multiple 'interesting' issue with this though:
1) If not defined, it will currently be enabled, but builtin, due to
   ``default HW_RANDOM``
2) Due to poor testing results of the HWRNG on several rk3566 based
   devices, the HWRNG is currently disabled in ``rk356x-base.dtsi``
   and only enabled in ``rk3568.dtsi``

ad 1) Due to a remark by Aurelien Jarno I looked further into this and
then notice that most have ``default HW_RANDOM`` and that enabling
``UML_RANDOM`` selects ``HW_RANDOM`` which in turn enables most
HW_RANDOM modules. This has now resulted in this patch set:
https://lore.kernel.org/linux-crypto/cover.1736946020.git.dsimic@manjaro.org/

It's better to explicitly enable it as module so that if the ``default``
line changes it will still be enabled and as it is now build as module,
it will only be loaded when the device has it enabled in the DeviceTree.

ad 2) Unfortunately the commit which disabled it in ``rk356x-base.dtsi``
did not document/link to the whole discussion that lead to it, but
there's now a patch which will fix that:
https://lore.kernel.org/linux-rockchip/6b272e2f8f916c04b05db50df621659a5a7f29ab.1733149874.git.dsimic@manjaro.org/

Note that this disables the HWRNG on rk3566 based devices *by default*,
but if testing shows that it does work properly on a certain device, the
dts for that device can be updated to enable the HWRNG for that device.