Significant changes to source package handling
Previously, we did not cope well with there being multiple versions of
source for a given package in the archive. If that set of source
packages did not all fit on a disc in one piece, we would fail the
build altogether. This caused issues in the case of exlive-extra (see

Whenever we generate a list of source packages now, we add a
package *per version* and handle them individually when adding /
removing them. This adds a little complexity, but not *too* much.

Interfaces have changed - all the paths to generating the list.mid /
packages / packages.$ARCH now append the version in the package info
line. We also now add the calculated size to the source package data
for safety and consistency in the data now. i.e.:

  source:main:gcc-14

now becomes

  source:main:gcc-14:94922561:14.2.0-1
  source:main:gcc-14:96397757:14.2.0-8
  source:main:gcc-14:96639410:14.2.0-9
  source:main:gcc-14:96811484:14.2.0-11
  source:main:gcc-14:96823306:14.2.0-12

(multiple lines now covering sizes and versions)

Binary packages don't need versions to be tracked (and it would make
things much harder to generate), so we use the string "<DEFAULT>" as a
placeholder there.

Tested on all of:

- amd64 DVD set
- source DVD set
- combined source and amd64 DVD set