Fix retrospectively wrongly generated debian/changelog
While we usually do not touch older entries this got wrong when I run
d-k-prerelease and the new upstream version should not be wrongly
associated. Fix it onetime up.

I suspect this happended because the maintainer trailer of the separate
stanza did not contain a space, but needs more investigation to avoid it
happends again.