Exclude firmware-marvell-prestera on all archs.
A very quick look into the firmware-marvell-prestera package suggests
the target devices are managed through the prestera.ko and
prestera_pci.ko modules, which are controlled by the CONFIG_PRESTERA
and CONFIG_PRESTERA_PCI options. The former is unset and the latter
doesn't appear at all (since it depends on the former) in trixie's
current kernel (as of 6.12.27-1).

Spotted via easy-build.sh, configured with NETINST and its historical
680M limitation (while official builds use STICK1GB instead). While
it shouldn't be needed for official builds, it makes sense not to
include something big (package size: 59M) that's not useful.