Fix sound firmware inclusion logic.
The intent seems to be detecting which initrd supports speech synthesis,
starting with the text installer, then falling back to the graphical
installer (as hinted by if-elif tests instead of two independent if
tests).

The logic is buggy on two accounts:
 - File presence is to be tested via -e or -f, -n is quite different.
 - Missing end quotes in both tests make the logic work entirely by
   accident: the quote in the if test gets terminated on the second
   line, and we end up with an overlong line alternating quoted and
   unquoted portions… The opening [ (if) is finally matched by the
   closing ] (elif), the overlong string in between has indeed a
   non-zero length, which makes the -n test return true, and we end
   up embedding the sound firmware in gtk/initrd.gz by pure luck.

Coinkydink level: 11/10.

Fix testing and quoting for all affected releases (bookworm and trixie).