Exclude more fonts udebs.
Fonts packages got revamped/renamed a number of releases ago but that
wasn't reflected in the exclude list, meaning wasting a lot of space.

All useful packages are expected to be used when graphical targets are
built, and there is no support for loading fonts at runtime anyway (as
seen while trying to compress some fonts for better Japanese support,
see #1037256 and relevant merge requests).

Link: https://lists.debian.org/debian-cd/2025/05/msg00016.html