Exclude wget-udeb and some of its dependencies.
It's been included in build/pkg-lists/base back in 2016, and currently
pulls both libcrypto3-udeb and libssl3-udeb, so let's exclude those as
well.

This should be fine for trixie but might break for forky, depending on
if and when SONAME bumps happen and which packages are rebuilt, so the
dependencies part of this commit should probably get reverted when the
forky release cycle starts.

Link: https://lists.debian.org/debian-cd/2025/05/msg00016.html